requirements.txt
# requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
google-generativeai>=0.8.3
protobuf>=5.27.0,<6.0.0
Werkzeug==2.3.7

